requirements.txt
langchain
faiss-cpu
sentence-transformers
fastapi
uvicorn
